:truck: Seperate mongoose related code with typescript type definition
