+++ packages/server/src/models/User.model.ts
@@ -1,58 +1,6 @@
 import mongoose, { Document, Schema } from 'mongoose'
 import 'reflect-metadata'
-import { Field, ID, ObjectType } from 'type-graphql'
-
-@ObjectType()
-export class TokenInfo {
-  @Field(() => String, { nullable: true })
-  token: string | null
-
-  @Field(() => Date, { nullable: true })
-  expires: Date | null
-
-  @Field(() => Boolean, { nullable: true })
-  isVerified?: boolean
-}
-
-@ObjectType()
-export class PurchasedInfo {
-  @Field()
-  isPurchased: boolean
-
-  @Field(() => Date, { nullable: true })
-  purchasedAt: Date | null
-
-  @Field(() => Date, { nullable: true })
-  expiresAt: Date | null
-}
-
-@ObjectType()
-export class User {
-  @Field(() => ID)
-  id: string
-
-  @Field()
-  email: string
-
-  @Field()
-  nickname: string
-
-  @Field(() => Date) // Specify Date type for createdAt
-  createdAt: Date
-
-  // Note: Typically, you wouldn't expose sensitive fields like password via GraphQL
-  // @Field()
-  // password: string;
-
-  @Field(() => TokenInfo)
-  emailVerification: TokenInfo
-
-  @Field(() => TokenInfo)
-  resetPassword: TokenInfo
-
-  @Field(() => PurchasedInfo)
-  purchasedInfo: PurchasedInfo
-}
+import { PurchasedInfo, TokenInfo } from '../types/user'
 
 const tokenInfoSchema: Schema = new Schema({
   token: { type: String, default: null },
